refactor: core module 의 이름을 common 으로 변경하라
