+++ module-common/src/test/java/kitchenpos/common/domain/NameTest.java
@@ -1,7 +1,7 @@
-package kitchenpos.core.domain;
+package kitchenpos.common.domain;
 
-import kitchenpos.core.exception.BadRequestException;
-import kitchenpos.core.exception.ExceptionType;
+import kitchenpos.common.exception.BadRequestException;
+import kitchenpos.common.exception.ExceptionType;
 import kitchenpos.domain.Name;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
