+++ module-common/src/test/resources/application.properties
