--- module-core/src/test/java/kitchenpos/ModuleCoreApplicationTests.java
@@ -1,11 +1,0 @@
-package kitchenpos;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-class ModuleCoreApplicationTests {
-    @Test
-    void contextLoads() {
-    }
-}
