+++ module-menu/src/test/java/kitchenpos/menu/application/MenuServiceTest.java
@@ -18,10 +18,10 @@ import java.math.BigDecimal;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
-import kitchenpos.core.exception.BadRequestException;
-import kitchenpos.core.exception.CannotCreateException;
-import kitchenpos.core.exception.ExceptionType;
-import kitchenpos.core.exception.NotFoundException;
+import kitchenpos.common.exception.BadRequestException;
+import kitchenpos.common.exception.CannotCreateException;
+import kitchenpos.common.exception.ExceptionType;
+import kitchenpos.common.exception.NotFoundException;
 import kitchenpos.menu.application.validator.MenuCreationValidator;
 import kitchenpos.menu.application.validator.MenuPriceValidator;
 import kitchenpos.menu.application.validator.MenuValidator;
