+++ module-order/src/main/java/kitchenpos/order/event/TableEmptyChangeHandler.java
@@ -1,8 +1,8 @@
 package kitchenpos.order.event;
 
 import java.util.Arrays;
-import kitchenpos.core.exception.CannotUpdateException;
-import kitchenpos.core.exception.ExceptionType;
+import kitchenpos.common.exception.CannotUpdateException;
+import kitchenpos.common.exception.ExceptionType;
 import kitchenpos.order.domain.OrderRepository;
 import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.table.domain.OrderTable;
