+++ module-order/src/test/java/kitchenpos/order/domain/OrderTest.java
@@ -6,8 +6,8 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.mockito.Mockito.mock;
 
 import java.util.Collections;
-import kitchenpos.core.exception.CannotUpdateException;
-import kitchenpos.core.exception.ExceptionType;
+import kitchenpos.common.exception.CannotUpdateException;
+import kitchenpos.common.exception.ExceptionType;
 import kitchenpos.order.application.OrderValidator;
 import kitchenpos.order.dto.request.OrderRequest;
 import org.junit.jupiter.api.BeforeEach;
