+++ module-product/src/test/java/kitchenpos/product/domain/ProductTest.java
@@ -1,7 +1,7 @@
 package kitchenpos.product.domain;
 
-import kitchenpos.core.exception.BadRequestException;
-import kitchenpos.core.exception.ExceptionType;
+import kitchenpos.common.exception.BadRequestException;
+import kitchenpos.common.exception.ExceptionType;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
