+++ module-table/src/main/java/kitchenpos/table/application/TableGroupValidator.java
@@ -1,8 +1,8 @@
 package kitchenpos.table.application;
 
-import kitchenpos.core.exception.BadRequestException;
-import kitchenpos.core.exception.CannotCreateException;
-import kitchenpos.core.exception.ExceptionType;
+import kitchenpos.common.exception.BadRequestException;
+import kitchenpos.common.exception.CannotCreateException;
+import kitchenpos.common.exception.ExceptionType;
 import kitchenpos.table.domain.OrderTable;
 import org.springframework.util.CollectionUtils;
