+++ module-table/src/test/java/kitchenpos/table/application/TableGroupValidatorTest.java
@@ -1,9 +1,8 @@
 package kitchenpos.table.application;
 
-import kitchenpos.core.exception.BadRequestException;
-import kitchenpos.core.exception.CannotCreateException;
-import kitchenpos.core.exception.ExceptionType;
-import kitchenpos.table.application.TableGroupValidator;
+import kitchenpos.common.exception.BadRequestException;
+import kitchenpos.common.exception.CannotCreateException;
+import kitchenpos.common.exception.ExceptionType;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.dto.request.OrderTableRequest;
 import org.junit.jupiter.api.BeforeEach;
