+++ settings.gradle
@@ -1,8 +1,8 @@
 rootProject.name = 'kitchenpos'
 
 include 'module-order'
-include 'module-core'
 include 'module-product'
 include 'module-menu'
 include 'module-table'
+include 'module-common'
